src/main/webapp/app/entities/biopsy/route/biopsy-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IBiopsy, Biopsy } from '../biopsy.model';
import { BiopsyService } from '../service/biopsy.service';

import { BiopsyRoutingResolveService } from './biopsy-routing-resolve.service';

describe('Biopsy routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: BiopsyRoutingResolveService;
  let service: BiopsyService;
  let resultBiopsy: IBiopsy | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(BiopsyRoutingResolveService);
    service = TestBed.inject(BiopsyService);
    resultBiopsy = undefined;
  });

  describe('resolve', () => {
    it('should return IBiopsy returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBiopsy = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBiopsy).toEqual({ id: 123 });
    });

    it('should return new IBiopsy if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBiopsy = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultBiopsy).toEqual(new Biopsy());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Biopsy })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultBiopsy = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultBiopsy).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
